# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform - Open API",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/openapi/v1",
    "paths": {
...
        "/knowledge/doc": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "删除文档",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "删除文档",
                "parameters": [
                    {
                        "description": "删除文档请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeleteDocReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/config": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取文档配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "获取文档配置信息",
                "parameters": [
                    {
                        "type": "string",
                        "name": "docId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "knowledgeId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.DocConfigResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/knowledge/doc/export": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "知识库文档导出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "知识库文档导出",
                "parameters": [
                    {
                        "description": "知识库文档导出请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.KnowledgeDocExportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/import": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "上传文档",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "上传文档",
                "parameters": [
                    {
                        "description": "文档上传请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DocImportReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/import/tip": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取知识库异步上传任务提示：有正在执行的异步上传任务/最近一次上传任务的失败信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "获取知识库异步上传任务提示",
                "parameters": [
                    {
                        "type": "string",
                        "name": "knowledgeId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/knowledge/doc/list": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "获取知识库文档列表，不展示状态为无效（-1）的文档数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "获取文档列表",
                "parameters": [
                    {
                        "description": "文档列表查询请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DocListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.DocPageResult"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/knowledge/doc/update/config": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "更新文档配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "更新文档配置",
                "parameters": [
                    {
                        "description": "更新文档配置请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DocConfigUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.DeleteDocReq": {
            "type": "object",
            "required": [
                "docIdList",
                "knowledgeId"
            ],
            "properties": {
                "docIdList": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "knowledgeId": {
                    "type": "string"
                }
            }
        },
...
        "request.DocConfigUpdateReq": {
            "type": "object",
            "required": [
                "docAnalyzer",
                "docIdList",
                "docSegment",
                "knowledgeId"
            ],
            "properties": {
                "asrModelId": {
                    "description": "asr模型id",
                    "type": "string"
                },
                "docAnalyzer": {
                    "description": "文档解析类型 text / ocr  / model / asr / multimodal",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docIdList": {
                    "description": "文档id列表，当更新配置时需要传",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docImportType": {
                    "description": "文档导入类型，0：文件上传，1：单条url上传，2.文件url上传",
                    "type": "integer"
                },
                "docMetaData": {
                    "description": "元数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.DocMetaData"
                    }
                },
                "docPreprocess": {
                    "description": "文本预处理规则 replaceSymbols / deleteLinks",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docSegment": {
                    "description": "文档分段配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.DocSegment"
                        }
                    ]
                },
                "knowledgeId": {
                    "description": "知识库id",
                    "type": "string"
                },
                "multimodalModelId": {
                    "description": "多模态模型id",
                    "type": "string"
                },
                "parserModelId": {
                    "description": "模型解析或ocr模型id",
                    "type": "string"
                }
            }
        },
        "request.DocImportReq": {
            "type": "object",
            "required": [
                "docAnalyzer",
                "docInfoList",
                "docSegment",
                "knowledgeId"
            ],
            "properties": {
                "asrModelId": {
                    "description": "asr模型id",
                    "type": "string"
                },
                "docAnalyzer": {
                    "description": "文档解析类型 text / ocr  / model / asr / multimodal",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docImportType": {
                    "description": "文档导入类型，0：文件上传，1：单条url上传，2.文件url上传",
                    "type": "integer"
                },
                "docInfoList": {
                    "description": "上传文档列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.DocInfo"
                    }
                },
                "docMetaData": {
                    "description": "元数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.DocMetaData"
                    }
                },
                "docPreprocess": {
                    "description": "文本预处理规则 replaceSymbols / deleteLinks",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docSegment": {
                    "description": "文档分段配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/request.DocSegment"
                        }
                    ]
                },
                "knowledgeId": {
                    "description": "知识库id",
                    "type": "string"
                },
                "multimodalModelId": {
                    "description": "多模态模型id",
                    "type": "string"
                },
                "parserModelId": {
                    "description": "模型解析或ocr模型id",
                    "type": "string"
                }
            }
        },
        "request.DocInfo": {
            "type": "object",
            "properties": {
                "docId": {
                    "description": "文档id",
                    "type": "string"
                },
                "docName": {
                    "description": "文档名称",
                    "type": "string"
                },
                "docSize": {
                    "description": "文档大小，单位字节(Byte)",
                    "type": "integer"
                },
                "docType": {
                    "description": "文档后缀",
                    "type": "string"
                },
                "docUrl": {
                    "description": "文档url（可选）",
                    "type": "string"
                }
            }
        },
        "request.DocListReq": {
            "type": "object",
            "required": [
                "knowledgeId",
                "pageSize"
            ],
            "properties": {
                "docIdList": {
                    "description": "文档id列表，只用于返回所选文档的集合，该值不为空时，其他筛选条件将被忽略",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docName": {
                    "type": "string"
                },
                "graphStatus": {
                    "description": "图谱状态：-1.全部 0.待处理 1.解析中 2.解析成功 3.解析失败",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "knowledgeId": {
                    "type": "string"
                },
                "metaValue": {
                    "type": "string"
                },
                "pageNo": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "status": {
                    "description": "文档状态：-1-全部， 0-待处理， 1- 处理完成， 2-正在审核中，3-正在解析中，4-审核未通过，5-解析失败",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "request.KnowledgeDocExportReq": {
            "type": "object",
            "required": [
                "knowledgeId"
            ],
            "properties": {
                "docIdList": {
                    "description": "文档id",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "knowledgeId": {
                    "description": "知识库id",
                    "type": "string"
                }
            }
        },
...
        "response.DocConfigResult": {
            "type": "object",
            "properties": {
                "asrModelId": {
                    "description": "asr模型id",
                    "type": "string"
                },
                "docAnalyzer": {
                    "description": "文档解析类型",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docImportType": {
                    "description": "文档导入类型，0：文件上传，1：url上传，2.批量url上传",
                    "type": "integer"
                },
                "docPreprocess": {
                    "description": "文本预处理规则",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "docSegment": {
                    "description": "分段信息配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.DocSegment"
                        }
                    ]
                },
                "multimodalModelId": {
                    "description": "多模态模型id",
                    "type": "string"
                },
                "parserModelId": {
                    "description": "ocr模型id",
                    "type": "string"
                }
            }
        },
...
        "response.DocPageResult": {
            "type": "object",
            "properties": {
                "docKnowledgeInfo": {
                    "$ref": "#/definitions/response.DocKnowledgeInfo"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ListDocResp"
                    }
                },
                "pageNo": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.DocSegment": {
            "type": "object",
            "required": [
                "segmentMethod"
            ],
            "properties": {
                "maxSplitter": {
                    "description": "可分隔最大值（只有自定义分段必填）",
                    "type": "integer"
                },
                "overlap": {
                    "description": "可重叠值（只有自定义分段必填）",
                    "type": "number"
                },
                "segmentMethod": {
                    "description": "分段方法 0：通用分段；1：父子分段",
                    "type": "string"
                },
                "segmentType": {
                    "description": "分段方式，只有通用分段必填 0：自动分段；1：自定义分段",
                    "type": "string"
                },
                "splitter": {
                    "description": "分隔符（只有自定义分段必填）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subMaxSplitter": {
                    "description": "可分隔最大值（只有父子分段必填）",
                    "type": "integer"
                },
                "subSplitter": {
                    "description": "分隔符（只有父子分段必填）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}